requirements.txt
colorama==0.3.9
PyYAML==3.12
requests==2.18.4
